[Fix](bangc-ops):fix large tensor bug.

+++ bangc-ops/core/tensor.h
@@ -71,8 +71,8 @@ struct mluOpTensorStruct {
 
   /* struct */
   int dim = 0;
-  int total_element_num = 0;
-  int total_tensor_size = 0;
+  uint64_t total_element_num = 0;
+  uint64_t total_tensor_size = 0;
   // if dimNb > MLUOP_DIM_MAX (8), using larger_dims, malloc it and dims point
   // it. else, using normal_dims, dont need malloc and free.
   int normal_dims[MLUOP_DIM_MAX] = {-1};
